% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tenx.R
\name{add_meta_data_cds}
\alias{add_meta_data_cds}
\title{Add colData (aka pData) to a cell_data_set}
\usage{
add_meta_data_cds(cds, meta, cds_col = "sample", meta_col = "sample")
}
\arguments{
\item{folders}{A vector of cellranger folders (full folder name is ideal).  These must contain an "outs" subfolders with
two files; 1) raw_feature_bc_matrix.h5, and 2) filtered_feature_bc_matrix.h5.  For aggregated samples
run with cellrangers 'aggregate' function, this folder should also contain the aggregation.csv file.}

\item{samplenames}{An optional vector that corresponds to the names you would like to give to
each element in filelist.  This defaults to the the basename of the filelist element i.e.}

\item{unfiltered}{This parameter, if true, returns a list containing 1) the filtered cds,
2) a list of unfiltered cds for each sample in 'folders'.}

\item{empty.droplet.threshold}{minimum number of umis per droplet to include in unfiltered output (default 15)}

\item{expressed_genes}{If true, this option removes genes that do not have expression in at least a 
minumum number of cells (cell_min parameter)}

\item{cell_min}{Minimum number of cells that a gene needs to be expressed for the gene to be included in the}

\item{aggregated}{will read a cellranger aggregate folder, this is currently only supported for 1 folder.  The 'aggregation.csv'
file must be included in the 'outs' folder.}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
This function will take a cds and add sample levels phenodata to all the cells from that
}
