% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backspin.R
\name{backspin}
\alias{backspin}
\title{Backspin}
\usage{
backspin(data, numLevels = 2, first_run_iters = 10,
  first_run_step = 0.05, runs_iters = 8, runs_step = 0.25,
  split_limit_g = 2, split_limit_c = 2, stop_const = 1.15,
  low_thrs = 0.2, verbose = T)
}
\arguments{
\item{data}{Rows should be genes and columns single cells/samples.}

\item{first_run_iters}{Number of the iterations used in the preparatory SPIN. Default 10}

\item{first_run_step}{the step parameter passed to _generate_widlist for the preparatory SPIN. Default 0.05}

\item{runs_iters}{the iterations parameter passed to the _divide_to_2and_resort. influences all the SPIN iterations except the first}

\item{verbose}{Print the extra details of what is happening}

\item{levels}{The number of nested splits that will be tried by the algorithm}

\item{runs_step:}{the step parameter passed to the _divide_to_2and_resort. influences all the SPIN iterations except the first}

\item{split_limit_g:}{If the number of specific genes in a subgroup is smaller than this number splitting of that subgrup is not allowed}

\item{split_limit_c:}{If the number cells in a subgroup is smaller than this number splitting of that subgrup is not allowed}

\item{stop_const:}{minimum score that a breaking point has to reach to be suitable for splitting}

\item{low_thrs:}{genes with average lower than this threshold are assigned to either of the splitting group reling on genes that are higly correlated with them}
}
\description{
The BackSPIN biclustering algorithm was developed by Amit Zeisel and is described in Zeisel et al. 
Cell types in the mouse cortex and hippocampus revealed by single-cell RNA-seq Science 2015 (PMID: 25700174, 
doi: 10.1126/science.aaa1934). Please cite this paper if you use the BackSPIN algorithm in your work.
Original MATLAB implementation by Amit Zeisel. This repo contains a standalone command-line version of 
BackSPIN, implemented in Python by Gioele La Manno.
}
\references{
Zeisel et. al. Science. 2015 Mar 6;347(6226):1138-42. doi: 10.1126/science.aaa1934. Epub 2015 Feb 19.
}
