% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_PCA}
\alias{cluster_PCA}
\title{Still working on this function
 @export}
\usage{
cluster_PCA(
  cds,
  method = "louvain",
  k = 20,
  dims = NULL,
  weight = F,
  num_iter = 1,
  resolution_parameter = NULL,
  random_seed = 2020,
  verbose = T,
  partition_q_value = 0.05
)
}
\description{
Still working on this function
 @export
}
