% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_features}
\alias{common_features}
\title{Finds common features in a list of cds objects}
\usage{
common_features(cds_list)
}
\arguments{
\item{cds_list}{Input cell_data_set object.}
}
\description{
Machine learning algorithms often require features to be the same across 
datasets.  Thisfunction finds common features between a list of cell data set objects and 
returns a list of cds's that have the same features.  Note that this function uses rownames 
of the 'fData' DataFrame to find the intersect of features common to all cds's
}
