% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_peak.R
\name{density_peak}
\alias{density_peak}
\title{Cluster cells using Density Peak algorithm}
\usage{
density_peak(
  cds,
  rho = NULL,
  delta = NULL,
  reduction_method = c("UMAP", "tSNE"),
  gaussian = T,
  pData_col = "Cluster"
)
}
\arguments{
\item{cds}{the cell_data_set upon which to perform this operation}

\item{rho}{The threshold of local density (rho) used to select the density peaks}

\item{delta}{The threshold of local distance (delta) used to select the density peaks}

\item{gaussian}{A logic flag passed to densityClust function in desnityClust package to determine whether or not Gaussian kernel will be used for calculating the local density}
}
\value{
an updated cell_data_set object
}
\description{
Unsupervised clustering of cells is a common step in many single-cell expression workflows. \
In an experiment containing a mixture of cell types, each cluster might correspond to a different \
cell type. This method takes a cell_data_set as input along with a requested number of clusters, \
clusters them using density peak clustering), and then returns the cell_data_set with the cluster assignments \
stored in the pData table as "Cluster". Use the plot_rho_delta to visualize the rho and delta parameters that will \
help determine the number of clusters to cluster.
}
\references{
Rodriguez, A., & Laio, A. (2014). Clustering by fast search and find of density peaks. \
Science, 344(6191), 1492-1496. doi:10.1126/science.1242072
}
