% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\name{doubletdetection}
\alias{doubletdetection}
\title{Doublet Detection}
\usage{
doubletdetection(cds, python_home = system("which python", intern =
  TRUE), module_file = paste(system.file(package = "m3addon"),
  "doubletdetection.py", sep = "/"), boost_rate = 0.25,
  n_components = 30, n_top_var_genes = 10000, use_phenograph = TRUE,
  n_iters = 25, normalizer = NULL, verbose = TRUE,
  standard_scaling = FALSE, p_thresh = 1e-07, voter_thresh = 0.9)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform Scrublet}

\item{python_home}{The python home directory where doubletdetection is installed}

\item{boost_rate}{(float, optional): Proportion of cell population size to 
produce as synthetic doublets.}

\item{n_components}{(int, optional): Number of principal components used for clustering.}

\item{n_top_var_genes}{(int, optional): Number of highest variance genes to use; other genes 
discarded. Will use all genes when zero. replace (bool, optional): If False, a cell will be 
selected as a synthetic doublet's parent no more than once.}

\item{use_phenograph}{(bool, optional): Set to False to disable PhenoGraph clustering
in exchange for louvain clustering implemented in scanpy. Defaults to True.}

\item{n_iters}{(int, optional): Number of fit operations from which to collect p-values. Defualt value is 25.}

\item{normalizer}{((sp_sparse) -> ndarray): Method to normalize raw_counts.Defaults to 
normalize_counts, included in this package. Note: To use normalize_counts with its 
pseudocount parameter changed from the default 0.1 value to some positive float `new_var`, 
use: normalizer=lambda counts: doubletdetection.normalize_counts(counts, pseudocount=new_var) 
random_state (int, optional): If provided, passed to PCA and used to seedrandom seed numpy's RNG. NOTE: 
PhenoGraph does not currently admit a random seed, and so this will not guarantee identical results 
across runs.}

\item{verbose}{(bool, optional): Set to False to silence all normal operation informational messages. Defaults to True.}

\item{standard_scaling}{(bool, optional): Set to True to enable standard scaling of normalized count matrix prior to PCA. 
Recommended when not using Phenograph. Defaults to False.}

\item{p_thresh}{(float, optional): hypergeometric test p-value threshold that determines per iteration doublet calls}

\item{voter_thresh}{(float, optional): fraction of iterations a cell must be called a doublet}
}
\value{
The input CellDataSet with an additional column added to pData with the doublet classification
}
\description{
See Gayoso, Adam, & Shor, Jonathan. (2018, July 17). 
DoubletDetection (Version v2.4). Zenodo. http://doi.org/10.5281/zenodo.2678042
}
\references{
Gayoso, Adam, & Shor, Jonathan. (2018, July 17). DoubletDetection (Version v2.4). Zenodo. http://doi.org/10.5281/zenodo.2678042
}
