% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{estimate_score}
\alias{estimate_score}
\title{Calculates geneset scores using totals}
\usage{
estimate_score(cds, marker_set1, fData_col = "gene_short_name")
}
\arguments{
\item{cds}{Input cell_data_set object.}

\item{marker_set1}{Vector of genes in the gene_metadata DataFrame (fData) that can be found in the column 'fData_col'}

\item{fData_col}{Character string denoting the gene_metadata DataFrame (fData) column that contains the genes in marker_set1.  Default = 'gene_short_name'}
}
\value{
Vector of single cell scores that are derived from the sum of all gene expression values in the geneset
that were size factor corrected and log-normalized.
}
\description{
Geneset scores are a score calculated for each single cell derived from more than one gene.  
In this implementation, the sum of the size-factor corrected, log-normalized gene expression for a give set of genes
is performed
}
