% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data using a set of features}
\usage{
extract_data(
  query,
  subject,
  annotation = "row.names",
  verbose = T,
  fill_output = T,
  duplicate_hits = "max.disp",
  ignore_strand = T
)
}
\arguments{
\item{query}{a vector of feature names or a GrangesObject}

\item{subject}{a SummarizedExperiment based object}

\item{annotation}{= A string specifying the target feature search for oin subject; Default is "row.names"}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can 
be set to FALSE for a cleaner output.}

\item{fill_output}{= A boolean indicating whether the function should return a subject object filled with zero data for 
features in query that were not found in the subject}

\item{duplicate_hits}{parameter for dealing with multiple subject hits in a range-based search.  Options are to select those
multiple hits based on the following: "max.mean", "max.var", "max.disp", "min.mean", "min.var", "min.disp". Default is "max.disp"}

\item{ignore_strand}{Whether to ignore strand for range-based searches}
}
\value{
a list cotaining 1) matrix of the extracted data, 2) the ratio of features
found in the subject to those present in the query, 3) those features not_found
}
\description{
This function extracts data from a subject object (a SummarizedExperiment-like 
object) using the features specified in query.  The function is parameterized to allow for finding features by either feature
name (annotation-based) or using GenomicRanges (range-based).  Annotation-based searching defaults
to the rownames of the subject, but this can be altered using annotation argument
which will search for these colnames in the subject's feature metadata.  The default behavior
of this function is to return a SummarizedExperiment object with the full feature-set of the query and thus
will add zeros for those features in the query not found in the subject. Set the full-output argument to TRUE
to modify this behavior
}
