% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{get_assay}
\alias{get_assay}
\title{get_assay helper}
\usage{
get_assay(se = NULL, assayName = NULL)
}
\description{
Adapted from: Jeffrey M. Granja, M. Ryan Corces, Sarah E. Pierce, S. Tansu Bagdatli, Hani Choudhry, Howard Y. Chang, William J. Greenleaf
doi: https://doi.org/10.1101/2020.04.28.066498
}
