% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GMT_to_list}
\alias{GMT_to_list}
\title{Convert a GMT File to a ragged list of genes}
\usage{
GMT_to_list(GMTfn)
}
\arguments{
\item{GMTfn}{GMT filename}
}
\description{
GMT Files (See MSigDB) are a convenient means of storing genesets
}
