% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{groupSums}
\alias{groupSums}
\title{Helper function for summing sparse matrix groups}
\usage{
groupSums(mat, groups = NULL, na.rm = TRUE, sparse = FALSE)
}
\description{
Helper function for summing sparse matrix groups
}
\references{
Granja, J. M.et al. (2019). Single-cell multiomic analysis identifies regulatory programs in mixed-phenotype 
acute leukemia. Nature Biotechnology, 37(12), 1458–1465.
}
