% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tenx.R
\name{load_cellranger_data_h5}
\alias{load_cellranger_data_h5}
\title{Make monocle3 cell_data_set using cellranger h5}
\usage{
load_cellranger_data_h5(
  folders = NULL,
  samplenames = NULL,
  unfiltered = F,
  files = NULL,
  empty.droplet.threshold = 15,
  expressed_genes = TRUE,
  cell_min = 1,
  aggregated = F,
  chemistry = "SC3Pv3",
  atac_feature = "peaks"
)
}
\arguments{
\item{folders}{A vector of cellranger folders (full folder name is ideal).  These must contain an "outs" subfolders with
two files; 1) raw_feature_bc_matrix.h5, and 2) filtered_feature_bc_matrix.h5.  For aggregated samples
run with cellrangers 'aggregate' function, this folder should also contain the aggregation.csv file.}

\item{samplenames}{An optional vector that corresponds to the names you would like to give to
each element in filelist.  This defaults to the the basename of the filelist element i.e.}

\item{unfiltered}{This parameter, if true, returns a list containing 1) the filtered cds,
2) a list of unfiltered cds for each sample in 'folders'.}

\item{files}{A vector of filtered h5 cellranger files.  Not compatible with unfiltered reading mode or aggregated mode.}

\item{empty.droplet.threshold}{minimum number of umis per droplet to include in unfiltered output (default 15)}

\item{expressed_genes}{If true, this option removes genes that do not have expression in at least a 
minumum number of cells (cell_min parameter)}

\item{cell_min}{Minimum number of cells that a gene needs to be expressed for the gene to be included in the}

\item{aggregated}{will read a cellranger aggregate folder, this is currently only supported for 1 folder.  The 'aggregation.csv'
file must be included in the 'outs' folder.}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
This function reads a vector of cellranger "out" folders for .h5 files.  This function
will create a cds of the data using cellranger thresholds for filtering out droplets that do not contain cells.
Optionally, this function can return the unfiltered data as well for further manipulation (ie adjustment of 
droplet thresholds) if desired.
}
