% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tenx.R
\name{load_STARsolo_data}
\alias{load_STARsolo_data}
\title{Make monocle3 cell_data_set using STARsolo}
\usage{
load_STARsolo_data(
  folders,
  samplenames = NULL,
  unfiltered = F,
  empty.droplet.threshold = 15,
  expressed_genes = TRUE,
  count_method = "Gene",
  cell_min = 1
)
}
\arguments{
\item{folders}{A vector of STARsolo folders (full folder name is ideal).  These must contain an "outs" subfolders with
three files: 1) features.tsv, 2) barcodes.tsv, and 3) matrix.mtx.}

\item{samplenames}{An optional vector that corresponds to the names you would like to give to
each element in filelist.}

\item{unfiltered}{This parameter, if true, returns a list containing 1) the filtered cds,
2) a list of unfiltered cds for each sample in 'folders'.}

\item{empty.droplet.threshold}{minimum number of umis per droplet to include in unfiltered output (default 15)}

\item{expressed_genes}{If true, this option removes genes that do not have expression in at least a 
minumum number of cells (cell_min parameter)}

\item{count_method}{(required) - A string denoting STARsolo output counting method.  (one of either: Gene, GeneFull, SJ, Velocyto)}

\item{cell_min}{Minimum number of cells that a gene needs to be expressed for the gene to be included in the}
}
\value{
a cell_data_set object or a list of items if unfiltered data is returned (see unfiltered)
}
\description{
This function reads a vector of STARsolo "out" folders for .mtx and .tsv files.  This function
will create a cds of the data using cellranger thresholds for filtering out droplets that do not contain cells.
Optionally, this function can return the unfiltered data as well for further manipulation (ie adjustment of 
droplet thresholds) if desired.
}
