% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soupX.R
\name{loadSoupX}
\alias{loadSoupX}
\title{Load Data for SoupX}
\usage{
loadSoupX(dataDir, cellIDs = NULL, channelName = NULL,
  readArgs = list(), includeFeatures = c("Gene Expression"),
  verbose = TRUE, method = c("SoupX_10X", "10X_H5", "STARsolo"), ...)
}
\arguments{
\item{dataDir}{Top level cellranger output directory (the directory that contains the "raw_gene_bc_matrices" folder).}

\item{cellIDs}{Barcodes of droplets that contain cells. If NULL, use the default cellranger set.}

\item{channelName}{The name of the channel to store. If NULL set to either names(dataDir) or dataDir is no name is set.}

\item{readArgs}{A list of extra parameters passed to Seurat::Read10X.}

\item{includeFeatures}{If multiple feature types are present, keep only the types mentioned here and collapse to a single matrix.}

\item{verbose}{Be verbose?}

\item{method}{type of input data; Use "SoupX_10X" for processing cellranger output, "10X_H5" for processing a folder of H5 
cellranger output files, and "STARsolo" for processing STARsolo output.}

\item{...}{Extra parameters passed to SoupChannel construction function.}
}
\value{
A SoupChannel object containing the count tables for the 10X dataset.
}
\description{
Loads unfiltered 10X data from each data-set and identifies which droplets are cells using the cellranger defaults.
}
