\name{m3addon-package}
\alias{m3addon-package}
\alias{m3addon}
\docType{package}
\title{
\packageTitle{m3addon}
}
\description{
\packageDescription{m3addon}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{m3addon}
\packageIndices{m3addon}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{m3addon}

Maintainer: \packageMaintainer{m3addon}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
