% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrect.R
\name{mnnCorrect_cds}
\alias{mnnCorrect_cds}
\title{Performs mNN batch correction on PCA data from a cds}
\usage{
mnnCorrect_cds(
  cds,
  batch_term,
  fast = TRUE,
  k = 20,
  sigma = 0.1,
  cos.norm.in = TRUE,
  cos.norm.out = TRUE,
  svd.dim = 0L,
  var.adj = TRUE,
  compute.angle = FALSE,
  subset.row = NULL,
  order = NULL,
  pc.approx = FALSE,
  irlba.args = list()
)
}
\arguments{
\item{cds}{Input cell_data_set object.}

\item{batch_term}{variable in ColData (pData) of cds that informs the batch to be corrected}

\item{fast}{Whether to perform fastMNN or mnnCorrect (see scran package)}

\item{min_expr}{Numeric indicating expression threshold}

\item{exprs_bin}{Boolean whether to bin genes by mean expression}

\item{exprs_cuts}{Numeic indicating number of bins if using exprs_bin}
}
\value{
Updated cell_data_set object
}
\description{

}
