% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_violin}
\alias{plot_genes_violin}
\title{Plot violin plot of gene expression by group}
\usage{
plot_genes_violin(cds_subset, grouping = "State", min_expr = NULL,
  scale_y = NULL, cell_size = 0.75, nrow = NULL, ncol = 1,
  panel_order = NULL, color_by = NULL, plot_trend = F,
  label_by_short_name = TRUE, relative_expr = TRUE, lognorm = FALSE,
  noise = FALSE, adjust = 1.3, scale = "width", jitter = FALSE,
  trim = FALSE, jitter_alpha = 0.3, round = FALSE,
  jitterzeros = FALSE, jitter_width = 0.05, jitter_height = 0,
  box_color = "red", violin_alpha = 1, boxplot = F, bwidth = 0.3,
  bcolor = "white")
}
\arguments{
\item{cds}{Input cell_data_set object.}
}
\description{
You can figure it out
}
