% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rho_delta}
\alias{plot_rho_delta}
\title{Plot rho and delta for Density Peak}
\usage{
plot_rho_delta(cds)
}
\arguments{
\item{cds}{Input cell_data_set object.}
}
\description{
Plots the decision map of density clusters.
}
