% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkProjection.R, R/bulk_projection.R
\name{project_bulk_data}
\alias{project_bulk_data}
\title{Project Bulk RNA-seq data into single cell subspace}
\usage{
project_bulk_data(
  cds = NULL,
  se = NULL,
  embedding_ncells = 5000,
  scale = F,
  reduced_dim = "LSI",
  embedding = "UMAP",
  features = c("annotation-based", "range-based"),
  n = 250,
  verbose = TRUE,
  threads = 6,
  force = F,
  binarize = F,
  seed = 2020
)

project_bulk_data(
  cds = NULL,
  se = NULL,
  embedding_ncells = 5000,
  scale = F,
  reduced_dim = "LSI",
  embedding = "UMAP",
  features = c("annotation-based", "range-based"),
  n = 250,
  verbose = TRUE,
  threads = 6,
  force = F,
  binarize = F,
  seed = 2020
)
}
\arguments{
\item{cds}{cell_data_set object}

\item{se}{a bulk Summarized Experiment.}

\item{reduced_dim}{A string specifying the reducedDim (currently LSI supported).}

\item{embedding}{A string specifying embedding type (UMAP supported).}

\item{n}{An integer specifying the number of subsampled "pseudo single cells" per bulk sample.}

\item{verbose}{A boolean value indicating whether to use verbose output during execution of this function. Can be set to FALSE for a cleaner output.}

\item{threads}{The number of threads used for parallel execution}
}
\description{
This function will Project Bulk RNA-seq data into single cell subspace. Adapted from: ArchR: An integrative and scalable software package for single-cell chromatin accessibility analysis
Jeffrey M. Granja, M. Ryan Corces, Sarah E. Pierce, S. Tansu Bagdatli, Hani Choudhry, Howard Y. Chang, William J. Greenleaf
doi: https://doi.org/10.1101/2020.04.28.066498

This function will Project bulk sequencing data into single cell subspace. Adapted from: ArchR: An integrative and scalable software package for single-cell chromatin accessibility analysis
Jeffrey M. Granja, M. Ryan Corces, Sarah E. Pierce, S. Tansu Bagdatli, Hani Choudhry, Howard Y. Chang, William J. Greenleaf
doi: https://doi.org/10.1101/2020.04.28.066498
}
