% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulking.R
\name{pseudobulk}
\alias{pseudobulk}
\title{Pseudobulk}
\usage{
pseudobulk(cds, by, n = 10)
}
\arguments{
\item{cds}{Input cell_data_set object.}

\item{by}{column colData from which to group pseudobulking efforts}

\item{n}{number of cells to bulk}
}
\value{
cds object
}
\description{
This function creates pseudocells comprosed of a composite of data from n cells using 
an additive method.
}
