% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plotting.R
\name{segment}
\alias{segment}
\title{Segment a Range}
\usage{
segment(range, fraction = 0.2, n = 1)
}
\arguments{
\item{cds}{Input cell_data_set object.}
}
\description{
This function takes a ordered range of 2 values, fractionates the range based 
on input fraction and returns the range of the nth fraction
}
