% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plotting.R
\name{sfc}
\alias{sfc}
\title{Produce nice colors (sf color)}
\usage{
sfc(n, scramble = F)
}
\arguments{
\item{n}{colors}

\item{scramble}{boolean whether to scramble the colors}
}
\description{
Nice colors.  Importantly this function brings two features that are useful.  1) scrambling the colors, 2) ideal color hue separation when n is low
}
