% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plotting.R
\name{split_plot}
\alias{split_plot}
\title{Plot a fixed size for plot and legend}
\usage{
split_plot(a.gplot, widths = c(4.5/6, 1.5/6))
}
\arguments{
\item{a.gplot}{A ggplot2 object}

\item{widths}{A vector of length 2 specifying the size of the plot grid and legend grid, respectively.}
}
\value{
grid arranged plot
}
\description{
Creates a grid object comprised of both a plot (ggplot only) and a legend.  Function takes
a ggplot object as input and returns a grid arrange object with plot on left, legend on right and sizes as dictated
by width argument
}
\references{
Puram, S. V. et al. Single-Cell Transcriptomic Analysis of Primary and 
Metastatic Tumor Ecosystems in Head and Neck Cancer. Cell 171, 1611.e1–1611.e24 (2017).
}
