% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulking.R
\name{ssgsea}
\alias{ssgsea}
\title{ssGSEA}
\usage{
ssgsea(X, gene_sets, alpha = 0.25, scale = T, norm = F, single = T)
}
\arguments{
\item{X}{matrix. Rows are genes. Columns are samples. Row names are symbols.}

\item{gene_sets}{list. Each element is a string vector with gene symbols.}

\item{alpha}{numeric. Parameter for ssGSEA, the default is 0.25}

\item{scale}{logical. If True, normalize the scores by number of genes in the gene sets.}

\item{norm}{logical. If True, normalize the scores by the absolute difference between max and min values.}

\item{single}{logical. If True, use ssGSEA algorithm, otherwise use GSEA.}
}
\value{
matrix containing enrichment scroes. Rows are gene sets, columns are samples.
}
\description{
This function runs ssGSEA as implemented here: https://gist.github.com/gaoce/39e0907146c752c127728ad74e123b33
}
\examples{
# Create a fake matrix
m = 100
n = 100
set.seed(1)
X = matrix(rnorm(m*n), m, n)
# Assign 'gene symbols' to row names
rownames(X) = 1:m
# Create 3 gene sets
gene_sets = list(a = sample(m, 5), b = sample(m, 5), c = sample(m, 5))
system.time(assign('a', GSVA::gsva(X, gene_sets, method = 'ssgsea')))
system.time(assign('b', ssgsea(X, gene_sets, scale = F, norm = T)))
identical(a, b)
}
