% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tf_idf_transform_v2}
\alias{tf_idf_transform_v2}
\title{Performs TF-IDF transformation on a cell_data_set v2}
\usage{
tf_idf_transform_v2(input)
}
\arguments{
\item{cds_list}{Input cell_data_set object or sparse matrix.}
}
\description{
Just like it sounds but different.
}
