% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimap.R
\name{trimap}
\alias{trimap}
\title{Dimensionality Reduction Using Triplet Constraints}
\usage{
trimap(
  cds,
  python_home = system("which python", intern = TRUE),
  num_dims = NULL,
  module_file = paste(system.file(package = "m3addon"), "trimap.py", sep = "/"),
  n_dims = 2,
  n_inliers = 10,
  n_outliers = 5,
  n_random = 5,
  distance = c("euclidean", "manhattan", "angular", "hamming"),
  lr = 1000,
  n_iters = 400,
  knn_tuple = NULL,
  apply_pca_trimap = FALSE,
  opt_method = c("dbd", "sd", "momentum"),
  verbose = TRUE,
  weight_adj = 500,
  return_seq = FALSE
)
}
\arguments{
\item{n_dims}{Number of dimensions of the embedding (default = 2)}

\item{data}{input data samples (rows), features (columns)}

\item{n_inliers:}{Number of inlier points for triplet constraints (default = 10)}

\item{n_outliers:}{Number of outlier points for triplet constraints (default = 5)}

\item{n_random:}{Number of random triplet constraints per point (default = 5)}

\item{distance:}{Distance measure ('euclidean' (default), 'manhattan', 'angular', 'hamming')}

\item{lr:}{Learning rate (default = 1000.0)}

\item{n_iters:}{Number of iterations (default = 400)}

\item{knn_tuple:}{Use the pre-computed nearest-neighbors information in form of a tuple (knn_nbrs, knn_distances) (default = None)}

\item{apply_pca:}{Apply PCA to reduce the dimensions to 100 if necessary before the nearest-neighbor calculation (default = True)}

\item{opt_method:}{Optimization method ('sd': steepest descent,  'momentum': GD with momentum, 
'dbd': GD with momentum delta-bar-delta (default))}

\item{verbose:}{Print the progress report (default = True)}

\item{weight_adj:}{Adjusting the weights using a non-linear transformation (default = 500.0)}

\item{return_seq:}{Return the sequence of maps recorded every 10 iterations (default = False)}
}
\description{
Find a low-dimensional repersentation of the data by satisfying the sampled triplet constraints from the high-dimensional features.
}
\references{
TriMap: Large-scale Dimensionality Reduction Using Triplets. E Amid, MK Warmuth - arXiv preprint arXiv:1910.00204, 2019 - arxiv.org
}
