% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletFinder.R
\name{doubletFinder_v3}
\alias{doubletFinder_v3}
\title{doubletFinder}
\usage{
cds <- doubletFinder(cds, PCs, pN = 0.25, pK, nExp, reuse.pANN = FALSE)
}
\arguments{
\item{cds}{Input cell_data_set object.}

\item{PCs}{Number of statistically-significant principal components (e.g., as estimated from PC 
elbow plot)}

\item{pN}{The number of generated artificial doublets, expressed as a proportion of the merged 
real-artificial data. Default is set to 0.25, based on observation that DoubletFinder performance is
 largely pN-invariant (see McGinnis, Murrow and Gartner 2019, Cell Systems).}

\item{pK}{The PC neighborhood size used to compute pANN, expressed as a proportion of the merged 
real-artificial data. No default is set, as pK should be adjusted for each scRNA-seq dataset. Optimal 
pK values can be determined using mean-variance-normalized bimodality coefficient.}

\item{nExp}{The total number of doublet predictions produced. This value can best be estimated 
from cell loading densities into the 10X/Drop-Seq device, and adjusted according to the estimated 
proportion of homotypic doublets.}

\item{reuse.pANN}{Metadata column name for previously-generated pANN results. Argument should be set to 
FALSE (default) for initial DoubletFinder runs. Enables fast adjusting of doublet predictions for 
different nExp.}
}
\description{
Core doublet prediction function of the DoubletFinder package. Generates artifical 
doublets from an existing, pre-processed cdsrat object. Real and artificial data are then merged 
and pre-processed using parameters utilized for the existing cdsrat object. PC distance matrix is 
then computed and used the measure the proportion of artificial nearest neighbors (pANN) for every 
real cell. pANN is then thresholded according to the number of expected doublets to generate final 
doublet predictions.
}
\references{
McGinnis, Murrow and Gartner 2019, Cell Systems; https://github.com/chris-mcginnis-ucsf/DoubletFinder
}
