% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{enrichmentPlot}
\alias{enrichmentPlot}
\title{GSEA and plot}
\usage{
enrichmentPlot(pathway, stats, gseaParam = 1, segment = F, rug = T,
  rug_color = "black", segment_color = "black", dot_color = "green",
  dot_enhance = NULL, dot_enhance_size = 2, dot_shape = 21,
  dot_enhance_alpha = 0.7, dot_size = 1, return_data = FALSE,
  print_plot = FALSE, return_plot = TRUE)
}
\arguments{
\item{pathway}{= vector of genes}

\item{stats}{= vectir of ranked gene stats (usually fold change or SNR) with names \
that contain the gene name}

\item{rug}{= whether to make a rug plot}

\item{dot_enhance}{character string denoting a color that enhances the dot appearance \
with another color}

\item{all_the_rest_of_them}{Should be self explanatory}
}
\description{
Performs GSEA (using fgsea) and returns a GSEA-style enrichment plot
}
\references{
fgsea package
}
