% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_violin}
\alias{plot_genes_violin}
\title{Plot violin plot of gene expression by group}
\usage{
plot_genes_violin(cds_subset, grouping = "Cluster", min_expr = 0.1,
  scale_y = NULL, cell_size = 0.75, nrow = NULL, ncol = 1,
  panel_order = NULL, color_by = NULL, remove_zeros = T,
  plot_trend = F, label_by_short_name = TRUE, relative_expr = TRUE,
  lognorm = TRUE, noise = FALSE, adjust = 2, scale = "width",
  jitter = FALSE, trim = FALSE, jitter_alpha = 0.3, round = FALSE,
  jitterzeros = FALSE, jitter_width = 0.05, jitter_height = 0,
  box_color = "red", violin_alpha = 1, boxplot = F, bwidth = 0.3,
  bcolor = "white")
}
\arguments{
\item{cds}{Input cell_data_set object.}
}
\description{
You can figure it out
}
