% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulking.R
\name{remove_duplicated_rows}
\alias{remove_duplicated_rows}
\title{Removes data (collapses) with duplicated names}
\usage{
remove_duplicated_rows(
  cds,
  fdata_col = "gene_short_name",
  unique_labels = "id"
)
}
\arguments{
\item{cds}{Input cell_data_set object.}

\item{fdata_col}{the non-unique labels (i.e. Gene Symbol)}

\item{id}{the unique labels (i.e ENSEMBL)}
}
\value{
cds collapsed to only include the duplicated fdata_col label with the highest varience compated
to all other identical fdata_col labels
}
\description{
Will collapse a cds based on duplicated labels that have the greatest variance
}
